requirements.txt
Pillow
djangorestframework
django-cors-headers
python-dotenv
djangorestframework-simplejwt